+++ test/__mocks__/cableCarRoute.ts
@@ -1,0 +1,6 @@
+export default {
+    webSocketURL: 'websocketUrl',
+    provider: {
+        createConsumer: jest.fn(),
+    },
+}