+++ test/cableCar.spec.ts
@@ -1,4 +1,4 @@
-import CableCar from '../src/cableCar'
+import CableCar, { CableCarActionFilter } from '../src/cableCar'
 import configureMockStore from 'redux-mock-store'
 
 import {
@@ -10,8 +10,10 @@ import {
     mockUnsubscribe,
 } from './__mocks__/actioncable'
 
-jest.mock('actioncable')
+// actioncable mock
+const consumer = mockCreateConsumer()
 
+// redux mocks
 const middlewares = []
 const mockStore = configureMockStore(middlewares)
 const store = mockStore({})
@@ -21,15 +23,20 @@ describe('CableCar', () => {
         store.clearActions()
     })
 
+    it('sets the consumer', () => {
+        const cc = new CableCar(consumer, store, 'channel', {})
+        expect(cc.consumer).toEqual(consumer)
+    })
     it('sets the channel', () => {
-        const cc = new CableCar(store, 'channel', {})
+        const cc = new CableCar(consumer, store, 'channel', {})
         expect(cc.channel).toEqual('channel')
     })
     it('sets option defaults', () => {
-        const cc = new CableCar(store, 'channel', {})
+        const cc = new CableCar(consumer, store, 'channel', {})
         expect(cc.options).toEqual({
             params: {},
             silent: false,
+            matchChannel: false,
         })
     })
     it('sets options', () => {
@@ -37,18 +44,19 @@ describe('CableCar', () => {
         let options = {
             params: { roomId: 1 },
             silent: true,
+            matchChannel: false,
             initialized: callback,
             connected: callback,
             disconnected: callback,
             received: callback,
             rejected: callback,
         }
-        const cc = new CableCar(store, 'channel', options)
+        const cc = new CableCar(consumer, store, 'channel', options)
         expect(cc.options).toEqual(options)
     })
 
     it('creates a subscription', () => {
-        const cc = new CableCar(store, 'channel', {})
+        const cc = new CableCar(consumer, store, 'channel', {})
         expect(mockCreate).toHaveBeenCalled()
         expect(cc.subscription).toEqual({
             perform: mockPerform,
@@ -57,32 +65,14 @@ describe('CableCar', () => {
         })
     })
 
-    it('creates a subscription w/ custom websocket url', () => {
-        new CableCar(store, 'channel', {
-            webSocketURL: 'ws://custom',
-        })
-        expect(mockCreateConsumer).toHaveBeenCalledWith('ws://custom')
-    })
-
-    it('creates a subscription w/ custom provider', () => {
-        let mockCreate2 = jest.fn()
-        let customProvider = {
-            createConsumer: jest.fn(() => ({
-                subscriptions: { create: mockCreate2 },
-            })),
-        }
-        new CableCar(store, 'channel', {
-            provider: customProvider,
-        })
-        expect(customProvider.createConsumer).toHaveBeenCalledWith(null)
-        expect(mockCreate2).toHaveBeenCalled()
-    })
-
     it('destroys', () => {
         let callback = jest.fn()
-        const cc = new CableCar(store, 'channel', {}, callback)
+        const cc = new CableCar(consumer, store, 'channel', {}, callback)
         cc.active = true
+        cc.connected = true
         cc.destroy()
+        expect(cc.active).toEqual(false)
+        expect(cc.connected).toEqual(false)
         cc.destroy()
         expect(callback).toHaveBeenCalledTimes(1)
         expect(mockUnsubscribe).toHaveBeenCalledTimes(1)
@@ -90,13 +80,15 @@ describe('CableCar', () => {
 
     it('registers #initialized', () => {
         let mockInit = jest.fn()
-        const cc = new CableCar(store, 'channel', {
+        const cc = new CableCar(consumer, store, 'channel', {
             initialized: mockInit,
         })
+        expect(cc.active).toEqual(false)
         mockChannels['channel'].initialized()
+        expect(cc.active).toEqual(true)
         expect(mockInit).toHaveBeenCalled()
         let expectedAction = {
-            meta: { __cablecar__: true, __cablecarChannel__: 'channel' },
+            meta: { __cablecar__: true, channel: 'channel' },
             payload: {},
             type: 'redux-cablecar/INIT',
         }
@@ -105,13 +97,13 @@ describe('CableCar', () => {
 
     it('registers #rejected', () => {
         let mockRejected = jest.fn()
-        const cc = new CableCar(store, 'channel', {
+        const cc = new CableCar(consumer, store, 'channel', {
             rejected: mockRejected,
         })
         mockChannels['channel'].rejected()
         expect(mockRejected).toHaveBeenCalled()
         let expectedAction = {
-            meta: { __cablecar__: true, __cablecarChannel__: 'channel' },
+            meta: { __cablecar__: true, channel: 'channel' },
             payload: {},
             type: 'redux-cablecar/REJECTED',
         }
@@ -121,24 +113,23 @@ describe('CableCar', () => {
     it('registers #connected/#disconnected', () => {
         let mockConnected = jest.fn()
         let mockDisconnected = jest.fn()
-        const cc = new CableCar(store, 'channel', {
+        const cc = new CableCar(consumer, store, 'channel', {
             connected: mockConnected,
             disconnected: mockDisconnected,
         })
         expect(cc.active).toEqual(false)
         mockChannels['channel'].connected()
         expect(mockConnected).toHaveBeenCalled()
-        expect(cc.active).toEqual(true)
         mockChannels['channel'].disconnected()
         expect(mockDisconnected).toHaveBeenCalled()
         expect(cc.active).toEqual(false)
         let expectedAction1 = {
-            meta: { __cablecar__: true, __cablecarChannel__: 'channel' },
+            meta: { __cablecar__: true, channel: 'channel' },
             payload: {},
             type: 'redux-cablecar/CONNECTED',
         }
         let expectedAction2 = {
-            meta: { __cablecar__: true, __cablecarChannel__: 'channel' },
+            meta: { __cablecar__: true, channel: 'channel' },
             payload: {},
             type: 'redux-cablecar/DISCONNECTED',
         }
@@ -146,14 +137,14 @@ describe('CableCar', () => {
     })
 
     it('registers #received', () => {
-        new CableCar(store, 'channel', {})
+        new CableCar(consumer, store, 'channel', {})
         mockChannels['channel'].received({
             type: 'server',
             meta: { a: 1 },
             payload: { test: 'ing' },
         })
         let expectedAction = {
-            meta: { a: 1, __cablecar__: true, __cablecarChannel__: 'channel' },
+            meta: { a: 1, __cablecar__: true, channel: 'channel' },
             payload: { test: 'ing' },
             type: 'server',
         }
@@ -161,19 +152,55 @@ describe('CableCar', () => {
     })
 
     it('#send actions', () => {
-        let cc = new CableCar(store, 'channel', {})
+        let cc = new CableCar(consumer, store, 'channel', {})
+        cc.active = true
+        cc.connected = true
         cc.send('action')
         expect(mockSend).toHaveBeenCalledWith('action')
     })
 
+    it('#send actions (error on disconnected)', () => {
+        let cc = new CableCar(consumer, store, 'channel', {})
+        cc.connected = false
+        expect(() => {
+            cc.send('action')
+        }).toThrowError()
+    })
+
+    it('#send actions (error on inactive)', () => {
+        let cc = new CableCar(consumer, store, 'channel', {})
+        cc.active = false
+        expect(() => {
+            cc.send('action')
+        }).toThrowError()
+    })
+
     it('#perform actions', () => {
-        let cc = new CableCar(store, 'channel', {})
+        let cc = new CableCar(consumer, store, 'channel', {})
+        cc.active = true
+        cc.connected = true
         cc.perform('m1', 'p1')
         expect(mockPerform).toHaveBeenCalledWith('m1', 'p1')
     })
 
+    it('#perform actions (error on disconnected)', () => {
+        let cc = new CableCar(consumer, store, 'channel', {})
+        cc.connected = false
+        expect(() => {
+            cc.perform('m1', 'p1')
+        }).toThrowError()
+    })
+
+    it('#perform actions (error on inactive', () => {
+        let cc = new CableCar(consumer, store, 'channel', {})
+        cc.active = false
+        expect(() => {
+            cc.perform('m1', 'p1')
+        }).toThrowError()
+    })
+
     it('ignores internal actions', () => {
-        let cc = new CableCar(store, 'channel', {})
+        let cc = new CableCar(consumer, store, 'channel', {})
         let ccAction = {
             type: 'testing',
             meta: { __cablecar__: true },
@@ -182,7 +209,7 @@ describe('CableCar', () => {
     })
 
     it('utilizes permitted action function (default)', () => {
-        let cc = new CableCar(store, 'channel', {})
+        let cc = new CableCar(consumer, store, 'channel', {})
         let action1 = {
             type: 'RAILS_testing',
             meta: {},
@@ -192,7 +219,7 @@ describe('CableCar', () => {
 
     it('utilizes permitted action function (custom)', () => {
         let mockPermit = jest.fn()
-        let cc = new CableCar(store, 'channel', {
+        let cc = new CableCar(consumer, store, 'channel', {
             permittedActions: mockPermit,
         })
         let action1 = { type: 'test', meta: {} }
@@ -201,12 +228,94 @@ describe('CableCar', () => {
     })
 
     it('silences dispatches w/ silent option', () => {
-        let cc = new CableCar(store, 'channel', {
+        let cc = new CableCar(consumer, store, 'channel', {
             silent: true,
         })
         mockChannels['channel'].initialized()
         mockChannels['channel'].connected()
         let action1 = { type: 'RAILS-test', meta: {} }
         expect(store.getActions()).toEqual([])
+    })
+
+    describe('#permitsAction', () => {
+        it('permits actions properly (empty string)', () => {
+            const car = new CableCar(consumer, store, 'channel2', {
+                permittedActions: '',
+            })
+            let action1 = { type: 'anything' }
+            expect(car.permitsAction(action1)).toEqual(true)
+        })
+
+        it('permits actions properly (prefix string)', () => {
+            const car = new CableCar(consumer, store, 'channel2', {
+                permittedActions: 'PRE',
+            })
+            let action1 = { type: 'PRE_act1' }
+            let action2 = { type: 'PRE_act2' }
+            let action3 = { type: 'NON_PRE_act3' }
+            expect(car.permitsAction(action1)).toEqual(true)
+            expect(car.permitsAction(action2)).toEqual(true)
+            expect(car.permitsAction(action3)).toEqual(false)
+        })
+
+        it('permits actions properly (list of strings/regexp)', () => {
+            const car = new CableCar(consumer, store, 'channel2', {
+                permittedActions: ['EITHER', 'OR', /^YES/],
+            })
+            let action1 = { type: 'EITHER_act1' }
+            let action2 = { type: 'OR_act2' }
+            let action3 = { type: 'NOT_act3' }
+            let action4 = { type: 'YES_act3' }
+            expect(car.permitsAction(action1)).toEqual(true)
+            expect(car.permitsAction(action2)).toEqual(true)
+            expect(car.permitsAction(action3)).toEqual(false)
+            expect(car.permitsAction(action4)).toEqual(true)
+        })
+
+        it('permits actions properly (RegExp)', () => {
+            const car = new CableCar(consumer, store, 'channel2', {
+                permittedActions: /^START.+FINISH$/,
+            })
+            let action1 = { type: 'START_act1_FINISH' }
+            let action2 = { type: 'START_act2' }
+            let action3 = { type: 'NOT_act3_FINISH' }
+            expect(car.permitsAction(action1)).toEqual(true)
+            expect(car.permitsAction(action2)).toEqual(false)
+            expect(car.permitsAction(action3)).toEqual(false)
+        })
+
+        it('permits actions properly (function)', () => {
+            const car = new CableCar(consumer, store, 'channel2', {
+                permittedActions: (a: CableCarActionFilter) =>
+                    a['payload'] % 7 === 5,
+            })
+            let action1 = { type: 'EXACT', payload: 12 }
+            let action2 = { type: 'EXACTly', payload: 10 }
+            let action3 = { type: '-EXACT' }
+            expect(car.permitsAction(action1)).toEqual(true)
+            expect(car.permitsAction(action2)).toEqual(false)
+            expect(car.permitsAction(action3)).toEqual(false)
+        })
+
+        it('permits actions properly (matchChannel)', () => {
+            const car = new CableCar(consumer, store, 'channelOne', {
+                matchChannel: true,
+            })
+            const car2 = new CableCar(consumer, store, 'channelTwo', {
+                matchChannel: true,
+            })
+            let action1 = {
+                type: 'RAILS/first',
+                meta: { channel: 'channelOne' },
+            }
+            let action2 = {
+                type: 'RAILS/third',
+                meta: { channel: 'channelTwo' },
+            }
+            expect(car.permitsAction(action1)).toEqual(true)
+            expect(car.permitsAction(action2)).toEqual(false)
+            expect(car2.permitsAction(action1)).toEqual(false)
+            expect(car2.permitsAction(action2)).toEqual(true)
+        })
     })
 })